Move streamerMode functionality to separate file
