+++ src/routes/w/[title]/+page.svelte
@@ -1,6 +1,5 @@
 <script>
 	/* eslint-disable svelte/no-at-html-tags */
-	import { onMount } from 'svelte';
 	import { SquarePenIcon, HistoryIcon } from '@lucide/svelte';
 	import { formatRelative } from 'date-fns';
 	import { enGB } from 'date-fns/locale';
@@ -15,10 +14,10 @@
 	import ToC from '$lib/components/ToC.svelte';
 	import RandomButton from '$lib/components/RandomButton.svelte';
 	import CacheBustButton from '$lib/components/CacheBustButton.svelte';
-	import Button from '$lib/components/Button.svelte';
 	import Link from '$lib/components/Link.svelte';
 	import { isSystem } from '$lib/utils/isSystem.js';
 	import { getImageCacheURL } from '$lib/utils/getImageCacheURL.js';
+	import Article from '$lib/components/Article.svelte';
 
 	let streamerMode = $state(page.data.session?.user?.userSettings?.streamerMode);
 	let showAuthors = $state(false);
@@ -98,81 +97,6 @@
 	function toggleAuthor() {
 		showAuthors = !showAuthors;
 	}
-
-	onMount(() => {
-		/**
-		 * @type {ResizeObserver[]}
-		 */
-		let observers = [];
-
-		if (!streamerMode) {
-			return;
-		}
-
-		const streamerModeItems = /** @type {HTMLElement[]}*/ ([
-			...document.querySelectorAll('.article-root img'),
-			...document.querySelectorAll('.article-root [data-lexical-youtube]'),
-			...document.querySelectorAll('.article-root [data-lexical-twitch]'),
-			...document.querySelectorAll('.article-root video'),
-		]);
-
-		for (let index = 0; index < streamerModeItems.length; index++) {
-			const item = streamerModeItems[index];
-
-			const parent = item.parentElement;
-
-			if (!parent) {
-				continue;
-			}
-
-			parent.classList.add('relative');
-
-			const overlay = document.createElement('div');
-			overlay.innerHTML = 'Show';
-
-			overlay.classList.add(
-				'absolute',
-				'flex',
-				'cursor-pointer',
-				'items-center',
-				'justify-center',
-				'text-white',
-				'bg-black/50',
-				'whitespace-nowrap',
-				'z-10'
-			);
-
-			overlay.style.width = `${item.offsetWidth + 2}px`;
-			overlay.style.height = `${item.offsetHeight + 2}px`;
-
-			const positionOverlay = () => {
-				overlay.style.left = `${item.offsetLeft + item.offsetWidth / 2}px`;
-				overlay.style.top = `${item.offsetTop + item.offsetHeight / 2}px`;
-				overlay.style.width = `${item.offsetWidth + 2}px`;
-				overlay.style.height = `${item.offsetHeight + 2}px`;
-				overlay.style.transform = 'translate(-50%, -50%)';
-			};
-
-			item.insertAdjacentElement('afterend', overlay);
-			positionOverlay();
-
-			const ro = new ResizeObserver(positionOverlay);
-			ro.observe(parent);
-			observers.push(ro);
-
-			const reveal = () => {
-				item.classList.add('streamer-mode-disable');
-				overlay.removeEventListener('click', reveal);
-				overlay.remove();
-			};
-
-			overlay.addEventListener('click', reveal);
-		}
-
-		return () => {
-			observers.forEach((ro) => ro.disconnect());
-		};
-	});
 </script>
 
 <svelte:head>
@@ -211,7 +135,7 @@
 	{/if}
 </svelte:head>
 
-<Container class={streamerMode ? 'streamer-mode' : ''}>
+<Container>
 	<article class="relative flex grow flex-col gap-4">
 		<RandomButton />
 
@@ -254,7 +178,7 @@
 							<strong class="text-4xl">{rawTitle}</strong>
 						</div>
 
-						{@html html}
+						<Article {html} />
 					</main>
 				</Box>
